DOC2-24. Extracted document read/write services
