+++ docusafe2-business-api/src/main/java/de/adorsys/docusafe2/business/api/access/DocumentReadService.java
@@ -1,0 +1,8 @@
+package de.adorsys.docusafe2.business.api.access;
+
+import de.adorsys.docusafe2.business.api.types.ReadRequest;
+
+public interface DocumentReadService {
+
+    void read(ReadRequest request);
+}
