+++ docusafe2-business-api/src/main/java/de/adorsys/docusafe2/business/api/access/DocumentWriteService.java
@@ -1,0 +1,10 @@
+package de.adorsys.docusafe2.business.api.access;
+
+import de.adorsys.docusafe2.business.api.types.WriteRequest;
+
+public interface DocumentWriteService {
+
+    void write(WriteRequest request);
+}
+
+
