+++ docusafe2-business-impl/src/main/java/de/adorsys/docusafe2/business/impl/document/CMSDocumentWriteService.java
@@ -1,20 +1,22 @@
 package de.adorsys.docusafe2.business.impl.document;
 
 import de.adorsys.dfs.connection.api.service.api.ExtendedStoreConnection;
+import de.adorsys.docusafe2.business.api.access.DocumentWriteService;
 import de.adorsys.docusafe2.business.api.dfs.DFSConnectionService;
-import de.adorsys.docusafe2.business.impl.document.dto.WriteRequest;
+import de.adorsys.docusafe2.business.api.types.WriteRequest;
 
 import javax.inject.Inject;
 
-public class DocumentWriteService {
+public class CMSDocumentWriteService implements DocumentWriteService {
 
     private final DFSConnectionService dfs;
 
     @Inject
-    public DocumentWriteService(DFSConnectionService dfs) {
+    public CMSDocumentWriteService(DFSConnectionService dfs) {
         this.dfs = dfs;
     }
 
+    @Override
     public void write(WriteRequest request) {
         ExtendedStoreConnection connection = dfs.obtain(request.getTo());
     }
